Rework Navier-Stokes system for multiprocess execution.
Loading the results with "np.loadtxt" takes a significant amount of
time. Generate each trajectory in a separate subprocess so parsing the
floats from text can be done in parallel too. Reduce log file
contention to avoid slowdowns or conflicts. Instead, send PolyFEM logs
to a tar file in the output directory.